pyproject.toml
[build-system]
requires = ["flit_core==3.12.0"]
build-backend = "flit_core.buildapi"

[project]
name = "custom_plugin"
version = "1.0.2"
description = "Provider package apache-airflow-providers-databricks for Apache Airflow"
readme = "README.md"
authors = [
    {name="Sunil Kunchoor", email="sunilkunchoor@gmail.com"},
]

requires-python = ">=3.10"

# The dependencies should be modified in place in the generated file.
# Any change in the dependencies is preserved when the file is regenerated
# Make sure to run ``prek update-providers-dependencies --all-files``
# After you modify the dependencies, and rebuild your Breeze CI image with ``breeze ci-image build``
dependencies = [
    "apache-airflow==2.11.0",
    "apache-airflow-providers-common-compat>=1.8.0",
    "apache-airflow-providers-common-sql>=1.27.0",
    "requests>=2.32.0,<3",
    "databricks-sql-connector>=4.0.0",
    "databricks-sqlalchemy>=1.0.2",
    "aiohttp>=3.9.2, <4",
    "mergedeep>=1.3.4",
    'pandas>=2.1.2; python_version <"3.13"',
    'pandas>=2.2.3; python_version >="3.13"',
    "pyarrow>=16.1.0; python_version < '3.13'",
    "pyarrow>=18.0.0; python_version >= '3.13'",
]

[project.optional-dependencies]
"sdk" = [
    "databricks-sdk==0.10.0",
]
"azure-identity" = [
    "azure-identity>=1.3.1",
]
"fab" = [
    "apache-airflow-providers-fab>=2.2.0; python_version < '3.13'"
]
"standard" = [
    "apache-airflow-providers-standard"
]
"openlineage" = [
    "apache-airflow-providers-openlineage>=2.3.0"
]

[dependency-groups]
dev = [
    "apache-airflow",
    "apache-airflow-task-sdk",
    "apache-airflow-devel-common",
    "apache-airflow-providers-common-compat",
    "apache-airflow-providers-common-sql",
    "apache-airflow-providers-openlineage",
    # Additional devel dependencies (do not remove this line and add extra development dependencies)
    "deltalake>=1.1.3",
    "apache-airflow-providers-fab>=2.2.0; python_version < '3.13'",
    "apache-airflow-providers-microsoft-azure",
    "apache-airflow-providers-common-sql[pandas,polars]",
    "apache-airflow-providers-fab",
]


[tool.uv.sources]
# These names must match the names as defined in the pyproject.toml of the workspace items,
# *not* the workspace folder paths
apache-airflow = {workspace = true}
apache-airflow-devel-common = {workspace = true}
apache-airflow-task-sdk = {workspace = true}
apache-airflow-providers-common-sql = {workspace = true}
apache-airflow-providers-standard = {workspace = true}

[project.urls]
"Documentation" = "https://airflow.apache.org/docs/apache-airflow-providers-databricks/7.8.0"
"Changelog" = "https://airflow.apache.org/docs/apache-airflow-providers-databricks/7.8.0/changelog.html"
"Bug Tracker" = "https://github.com/apache/airflow/issues"
"Source Code" = "https://github.com/apache/airflow"
"Slack Chat" = "https://s.apache.org/airflow-slack"
"Mastodon" = "https://fosstodon.org/@airflow"
"YouTube" = "https://www.youtube.com/channel/UCSXwxpWZQ7XZ1WL3wqevChA/"

[project.entry-points."airflow.plugins"]
databricks_workflow_custom = "custom_plugin.databricks.plugins.databricks_workflow:DatabricksWorkflowPlugin"

[tool.flit.module]
name = "custom_plugin.databricks"